moving methods to common
